# back-end/src/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Power Socket API
  description: API for Smart Power Socket
  version: 1.0.0
servers:
  - url: http://localhost:3001
tags:
  - name: socket
    description: Smart socket
paths:
  /socket/{socketId}/charging-session:
    post:
      tags:
        - socket
      summary: Create new charging session for smart socket
      description: Start new charging session
      operationId: createSocketChargingSession
      parameters:
        - name: socketId
          in: path
          description: Socket ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        "$ref": "#/components/schemas/ChargingSessionRequest"
      responses:
        "200":
          description: Charging session created
        "400":
          description: Invalid input
      x-swagger-router-controller: socket
    delete:
      tags:
        - socket
      summary: Close existing charging session for smart socket
      description: Close existing charging session
      operationId: closeSocketChargingSession
      parameters:
        - name: socketId
          in: path
          description: Socket ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Charging session is closed
        "400":
          description: Invalid input
      x-swagger-router-controller: socket
    put:
      tags:
        - socket
      summary: Update existing charging session for smart socket
      description: Update existing charging session
      operationId: updateSocketChargingSession
      parameters:
        - name: socketId
          in: path
          description: Socket ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        "$ref": "#/components/schemas/ChargingSessionDetails"
      responses:
        "200":
          description: Charging session is updated
        "400":
          description: Invalid input
      x-swagger-router-controller: socket
components:
  schemas:
    ChargingSessionDetails:
      type: object
      properties:
        whConsumed:
          type: number
          format: double
      required:
        - "whConsumed"
    ChargingSessionRequest:
      type: object
      properties:
        paymentMethodId:
          type: string
      required:
        - "paymentMethodId"